Revert "Add index in extrinsics field to metadata from chain."
This reverts commit 2748878c3e642b12f6c417747002fe6c3c4bc847.
