+++ model/storage/chain/chain_model_metadata_store.py
@@ -18,58 +18,12 @@ class ChainModelMetadataStore(ModelMetadataStore):
         subtensor: bt.subtensor,
         wallet: Optional[bt.wallet] = None,
         subnet_uid: int = constants.SUBNET_UID,
-        archive_subtensor: bt.subtensor = bt.subtensor(network="archive"),
     ):
         self.subtensor = subtensor
-        # Use an 'archive' to ensure we can get extrinsics from older blocks.
-        self.archive_subtensor = archive_subtensor
         self.wallet = (
             wallet  # Wallet is only needed to write to the chain, not to read.
         )
         self.subnet_uid = subnet_uid
-
-    def _get_index_in_extrinsics_impl(self, hotkey: str, block: int) -> Optional[int]:
-        """Impl to get the index in extrinsics of the set commitment for the given block and hotkey.
-
-        Args:
-            hotkey (str): Hotkey that did the set commitment.
-            block (int): Block that the set commitment occurred in.
-        """
-        block_data = self.archive_subtensor.substrate.get_block(None, block)
-
-        if not block_data:
-            return None
-
-        # Check each extrinsic in the block for the set_commitment by the provided hotkey.
-        # Hotkeys can only set_commitment once every 20 minutes so just take the first we see.
-        for idx, extrinsic in enumerate(block_data["extrinsics"]):
-            # Check function name first, otherwise it may not have an address.
-            if (
-                extrinsic["call"]["call_function"]["name"] == "set_commitment"
-                and extrinsic["address"] == hotkey
-            ):
-                return idx
-
-        # This should never happen since we already confirmed there was metadata for this block.
-        bt.logging.trace(
-            f"Did not find any set_commitment for block {block} by hotkey {hotkey}"
-        )
-        return None
-
-    async def _get_index_in_extrinsics(self, hotkey: str, block: int) -> Optional[int]:
-        """Gets the index in extrinsics of the set commitment for the given block and hotkey.
-
-        Args:
-            hotkey (str): Hotkey that did the set commitment.
-            block (int): Block that the set commitment occurred in.
-        """
-        # Wrap calls to the subtensor in a subprocess with a timeout to handle potential hangs.
-        partial = functools.partial(
-            self._get_index_in_extrinsics_impl,
-            hotkey,
-            block,
-        )
-        return utils.run_in_subprocess(partial, 60)
 
     async def store_model_metadata(self, hotkey: str, model_id: ModelId):
         """Stores model metadata on this subnet for a specific wallet."""
@@ -98,11 +52,6 @@ class ChainModelMetadataStore(ModelMetadataStore):
         if not metadata:
             return None
 
-        extrinsic_index = await self._get_index_in_extrinsics(hotkey, metadata["block"])
-
-        if not extrinsic_index:
-            return None
-
         commitment = metadata["info"]["fields"][0]
         hex_data = commitment[list(commitment.keys())[0]][2:]
 
@@ -119,8 +68,100 @@ class ChainModelMetadataStore(ModelMetadataStore):
             )
             return None
 
-        model_metadata = ModelMetadata(
-            id=model_id, block=metadata["block"], extrinisic_index=extrinsic_index
-        )
+        model_metadata = ModelMetadata(id=model_id, block=metadata["block"])
 
         return model_metadata
+
+
+# Can only commit data every ~20 minutes.
+async def test_store_model_metadata():
+    """Verifies that the ChainModelMetadataStore can store data on the chain."""
+    model_id = ModelId(
+        namespace="TestPath", name="TestModel", hash="TestHash1", commit="1.0"
+    )
+
+    # Use a different subnet that does not leverage chain storage to avoid conflicts.
+    # TODO switch to a mocked version when it supports commits.
+    subtensor = bt.subtensor()
+
+    # Uses .env configured wallet/hotkey/uid for the test.
+    coldkey = os.getenv("TEST_COLDKEY")
+    hotkey = os.getenv("TEST_HOTKEY")
+    net_uid = int(os.getenv("TEST_SUBNET_UID"))
+
+    wallet = bt.wallet(name=coldkey, hotkey=hotkey)
+
+    metadata_store = ChainModelMetadataStore(
+        subtensor=subtensor, wallet=wallet, subnet_uid=net_uid
+    )
+
+    # Store the metadata on chain.
+    await metadata_store.store_model_metadata(hotkey=hotkey, model_id=model_id)
+
+    print(f"Finished storing {model_id} on the chain.")
+
+
+async def test_retrieve_model_metadata():
+    """Verifies that the ChainModelMetadataStore can retrieve data from the chain."""
+    expected_model_id = ModelId(
+        namespace="TestPath", name="TestModel", hash="TestHash1", commit="1.0"
+    )
+
+    # Use a different subnet that does not leverage chain storage to avoid conflicts.
+    # TODO switch to a mocked version when it supports commits.
+    subtensor = bt.subtensor()
+
+    # Uses .env configured hotkey/uid for the test.
+    net_uid = int(os.getenv("TEST_SUBNET_UID"))
+    hotkey_address = os.getenv("TEST_HOTKEY_ADDRESS")
+
+    # Do not require a wallet for retrieving data.
+    metadata_store = ChainModelMetadataStore(
+        subtensor=subtensor, wallet=None, subnet_uid=net_uid
+    )
+
+    # Retrieve the metadata from the chain.
+    model_metadata = await metadata_store.retrieve_model_metadata(hotkey_address)
+
+    print(f"Expecting matching model id: {expected_model_id == model_metadata.id}")
+
+
+# Can only commit data every ~20 minutes.
+async def test_roundtrip_model_metadata():
+    """Verifies that the ChainModelMetadataStore can roundtrip data on the chain."""
+    model_id = ModelId(
+        namespace="TestPath", name="TestModel", hash="TestHash1", commit="1.0"
+    )
+
+    # Use a different subnet that does not leverage chain storage to avoid conflicts.
+    # TODO switch to a mocked version when it supports commits.
+    subtensor = bt.subtensor()
+
+    # Uses .env configured wallet/hotkey/uid for the test.
+    coldkey = os.getenv("TEST_COLDKEY")
+    hotkey = os.getenv("TEST_HOTKEY")
+    net_uid = int(os.getenv("TEST_SUBNET_UID"))
+
+    wallet = bt.wallet(name=coldkey, hotkey=hotkey)
+
+    metadata_store = ChainModelMetadataStore(
+        subtensor=subtensor, wallet=wallet, subnet_uid=net_uid
+    )
+
+    # Store the metadata on chain.
+    await metadata_store.store_model_metadata(hotkey=hotkey, model_id=model_id)
+
+    # May need to use the underlying publish_metadata function with wait_for_inclusion: True to pass here.
+    # Otherwise it defaults to False and we only wait for finalization not necessarily inclusion.
+
+    # Retrieve the metadata from the chain.
+    model_metadata = await metadata_store.retrieve_model_metadata(hotkey)
+
+    print(f"Expecting matching metadata: {model_id == model_metadata.id}")
+
+
+if __name__ == "__main__":
+    # Can only commit data every ~20 minutes.
+    # asyncio.run(test_roundtrip_model_metadata())
+    # asyncio.run(test_store_model_metadata())
+    asyncio.run(test_retrieve_model_metadata())
